requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy>=2.0.47
psycopg[binary]>=3.2.0
jinja2==3.1.5
python-multipart==0.0.18
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
tenacity==9.0.0
python-dotenv>=1.0.0
